refactor: consolidate info services under Services/Info namespace
Moves existing information retrieval services and their interfaces into a new `Services/Info` directory and namespace. This improves overall project structure and prepares for the introduction of other service categories, such as `Services/Switch`.
