--- OrangePi.Display.Status.Service/InfoServices/IInfoService.cs
@@ -1,15 +1,0 @@
-﻿using OrangePi.Display.Status.Service.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace OrangePi.Display.Status.Service.InfoServices
-{
-    public interface IInfoService : IDisplayInfoService
-    {
-        Task<StatusValue> GetValue();
-        string Label { get; }
-    }
-}
