+++ OrangePi.Display.Status.Service/OrangePi.Display.Status.Service.csproj
@@ -9,7 +9,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="InfoServices\PiHoleInfoService.cs" />
+    <Compile Remove="Services\Info\PiHoleInfoService.cs" />
   </ItemGroup>
 
   <ItemGroup>
@@ -30,5 +30,9 @@
     <None Update="notify.mp3">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Services\Switch\" />
   </ItemGroup>
 </Project>
