+++ OrangePi.Display.Status.Service/Services/Info/HostInfoService.cs
@@ -2,7 +2,7 @@
 using OrangePi.Common.Services;
 using OrangePi.Display.Status.Service.Extensions;
 
-namespace OrangePi.Display.Status.Service.InfoServices
+namespace OrangePi.Display.Status.Service.Services.Info
 {
     public class HostInfoService : IHostInfoService
     {
