+++ OrangePi.Display.Status.Service/Services/Info/IDisplayInfoService.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace OrangePi.Display.Status.Service.InfoServices
+namespace OrangePi.Display.Status.Service.Services.Info
 {
     public interface IDisplayInfoService
     {
