+++ OrangePi.Display.Status.Service/Services/Info/IInfoService.cs
@@ -1,0 +1,15 @@
+﻿using OrangePi.Display.Status.Service.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OrangePi.Display.Status.Service.Services.Info
+{
+    public interface IInfoService : IDisplayInfoService
+    {
+        Task<StatusValue> GetValue();
+        string Label { get; }
+    }
+}
